pyproject.toml
[tool.black]
line-length = 88

[tool.ruff]
select = ["E", "F"]
line-length = 100

[tool.poetry]
name = "commdem-warriors-django"
version = "0.1.0"
description = "commdem warriors django backend application"
authors = ["anirudhflutter <anirudhflutter@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.16"
aniso8601 = "^9.0.1"
anyio = "^3.6.2"
asgiref = "^3.5.2"
async-timeout = "^4.0.2"
CacheControl = "^0.12.11"
cachetools = "^5.2.0"
certifi = "^2022.9.24"
cffi = "^1.15.1"
charset-normalizer = "*"
cryptography = "^38.0.3"
Deprecated = "^1.2.13"
dj-database-url = "^1.0.0"
Django = "^4.1.2"
django-heroku = "^0.3.1"
django-phonenumber-field = "^7.0.0"
django-phonenumbers = "^1.0.1"
django-redis = "^5.2.0"
djangorestframework = "*"
fcm-django = "^1.0.12"
firebase-admin = "*"
google-api-core = "^2.10.2"
google-api-python-client = "^2.65.0"
google-auth = "^2.14.0"
google-auth-httplib2 = "^0.1.0"
google-cloud-core = "^2.3.2"
google-cloud-firestore = "^2.7.2"
google-cloud-storage = "^2.5.0"
google-crc32c = "*"
google-resumable-media = "^2.4.0"
googleapis-common-protos = "^1.56.4"
graphene = "^3.1.1"
graphene-django = "^3.0.0"
graphql-core = "^3.2.3"
graphql-relay = "*"
grpcio = "^1.50.0"
grpcio-status = "^1.50.0"
gunicorn = "*"
h11 = "*"
httpcore = "*"
httplib2 = "*"
httpx = "*"
idna = "^3.4"
msgpack = "^1.0.4"
onesignal-python-api = "^1.0.1"
onesignal-sdk = "*"
packaging = "*"
phonenumbers = "^8.12.57"
promise = "^2.3"
proto-plus = "^1.22.1"
protobuf = "^4.21.9"
psycopg2-binary = "^2.9.4"
pyasn1 = "^0.4.8"
pyasn1-modules = "^0.2.8"
pycparser = "^2.21"
PyJWT = "^2.6.0"
pylibmc = "*"
pyngrok = "^5.1.0"
pyparsing = "^3.0.9"
python-dateutil = "^2.8.2"
pytz = "^2022.5"
PyYAML = "^6.0"
redis = "^4.3.4"
requests = "^2.28.1"
rfc3986 = "^1.5.0"
rsa = "*"
six = "^1.16.0"
sniffio = "*"
sqlparse = "^0.4.3"
text-unidecode = "^1.3"
uritemplate = "*"
urllib3 = "^1.26.12"
whitenoise = "^6.2.0"
wrapt = "^1.14.1"
razorpay = "^1.3.0"
celery = "^5.2.7"
python-dotenv = "*"
gTTS = "*"
channels = "*"
googletrans = "*"
twilio = "*"
PyLint = "*"
pylint-django = "*"
black = "*"
ruff = "*"
isort = "*"
mypy = "^1.7.1" 
flake8 = "^6.1.0"
ffmpeg = "^1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
